[pci] Reduce size of ATS structure elements
Message-id: <20160214182735.8391.61376.stgit@amt.stowe>
Patchwork-id: 134326
O-Subject: [RHEL7.3 PATCH 22/66] PCI: Reduce size of ATS structure elements
Bugzilla: 1050021
RH-Acked-by: Stefan Assmann <sassmann@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Laine Stump <laine@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Jeremy McNicoll <jmcnicol@redhat.com>

Bugzilla: https://bugzilla.redhat.com/show_bug.cgi?id=1050021
Upstream Status: 67930995d7fb8ae7d2078822b563010b289ace2e


Conflicts:
  Had to adjust for kABI (see 'struct pci_dev' ATS related content).

commit 67930995d7fb8ae7d2078822b563010b289ace2e
Author: Bjorn Helgaas <bhelgaas@google.com>
Date:   Fri Jul 17 15:27:34 2015 -0500

    PCI: Reduce size of ATS structure elements

    The extended capabilities list is linked with 12-bit pointers, and the ATS
    Smallest Translation Unit and Invalidate Queue Depth fields are both 5
    bits.

    Use u16 and u8 to hold the extended capability address and the stu and qdep
    values.  No functional change.

+++ include/linux/pci.h
@@ -403,9 +403,9 @@ struct pci_dev_rh {
 #endif
 	RH_KABI_EXTEND(unsigned int ats_enabled:1) /* Address Translation Svc */
 #ifdef CONFIG_PCI_ATS
-	RH_KABI_EXTEND(int ats_cap)	/* ATS Capability offset */
-	RH_KABI_EXTEND(int ats_stu)	/* ATS Smallest Translation Unit */
-	RH_KABI_EXTEND(int ats_qdep)	/* ATs Invalidate Queue Depth */
+	RH_KABI_EXTEND(u16 ats_cap)	/* ATS Capability offset */
+	RH_KABI_EXTEND(u8 ats_stu)	/* ATS Smallest Translation Unit */
+	RH_KABI_EXTEND(u8 ats_qdep)	/* ATs Invalidate Queue Depth */
 	RH_KABI_EXTEND(atomic_t ats_ref_cnt)  /* num of VFs with ATS enabled */
 #endif
 };
